… Cargo.toml
… ecmascript/parser/Cargo.toml
… ecmascript/parser/src/lexer/util.rs
… ecmascript/parser/src/parser/expr/ops.rs
… ecmascript/parser/src/parser/input.rs
… ecmascript/parser/src/parser/macros.rs
… ecmascript/parser/src/parser/stmt.rs
… ecmascript/transforms/Cargo.toml
… ecmascript/transforms/benches/assets/AjaxObservable.ts
… ecmascript/transforms/benches/base.rs
… ecmascript/transforms/benches/basic.rs
… ecmascript/transforms/benches/compat.rs
… ecmascript/transforms/macros/Cargo.toml
… ecmascript/transforms/macros/src/fast.rs
… ecmascript/transforms/macros/src/lib.rs
… ecmascript/transforms/src/compat.rs
… ecmascript/transforms/src/compat/es2015/arrow.rs
… ecmascript/transforms/src/compat/es2016/exponentation.rs
… ecmascript/transforms/src/compat/es2017/async_to_generator.rs
… ecmascript/transforms/src/compat/es2018/object_rest_spread.rs
… ecmascript/transforms/src/compat/es2018/optional_catch_binding.rs
… ecmascript/transforms/src/compat/es2020.rs
… ecmascript/transforms/src/compat/es2020/class_properties.rs
… ecmascript/transforms/src/compat/es2020/nullish_coalescing/mod.rs
… ecmascript/transforms/src/compat/es2020/opt_chaining.rs
… ecmascript/transforms/src/compat/reserved_words.rs
… ecmascript/transforms/src/ext.rs
… ecmascript/transforms/src/lib.rs
… ecmascript/transforms/src/perf.rs
… ecmascript/transforms/src/typescript/strip.rs