---


### Initial benchmark result:

```
test result: ok. 0 passed; 0 failed; 2 ignored; 0 measured; 0 filtered out

     Running target/release/deps/es2016-61776247b3e61f44

running 1 test
test es2016 ... bench:   4,010,261 ns/iter (+/- 124,429)

test result: ok. 0 passed; 0 failed; 0 ignored; 1 measured; 0 filtered out

```

### After no_inline optimization in the parser

```
test result: ok. 0 passed; 0 failed; 2 ignored; 0 measured; 0 filtered out

     Running target/release/deps/es2016-cceb1a6e4b83b366

running 1 test
test es2016 ... bench:   3,804,603 ns/iter (+/- 352,097)

test result: ok. 0 passed; 0 failed; 0 ignored; 1 measured; 0 filtered out

```

### After optimizing parser slightly

```
test result: ok. 0 passed; 0 failed; 2 ignored; 0 measured; 0 filtered out

     Running target/release/deps/es2016-cceb1a6e4b83b366

running 1 test
test es2016 ... bench:   3,780,166 ns/iter (+/- 147,297)

test result: ok. 0 passed; 0 failed; 0 ignored; 1 measured; 0 filtered out

```

### Macro -> method

```
test result: ok. 0 passed; 0 failed; 2 ignored; 0 measured; 0 filtered out

     Running target/release/deps/es2016-cceb1a6e4b83b366

running 1 test
test es2016 ... bench:   3,753,983 ns/iter (+/- 126,315)

test result: ok. 0 passed; 0 failed; 0 ignored; 1 measured; 0 filtered out
```

### After migrating typescript::strip

Note: I renamed es2016 to es2016_cjs, added some more benchmarks, disabled lto.

```
test result: ok. 0 passed; 0 failed; 2 ignored; 0 measured; 0 filtered out

     Running target/release/deps/typescript-0a87800f9f443513

running 2 tests
test es2016     ... bench:   3,465,284 ns/iter (+/- 98,719)
test es2016_cjs ... bench:   3,486,083 ns/iter (+/- 100,640)

test result: ok. 0 passed; 0 failed; 0 ignored; 2 measured; 0 filtered out
```

### Added more benchmarks

```
running 17 tests
test base_parser ... bench:     936,720 ns/iter (+/- 53,895)
test es2015      ... bench:   5,651,283 ns/iter (+/- 2,093,096)
test es2015_cjs  ... bench:   5,364,132 ns/iter (+/- 153,060)
test es2016      ... bench:   3,505,975 ns/iter (+/- 86,921)
test es2016_cjs  ... bench:   3,580,391 ns/iter (+/- 176,638)
test es2017      ... bench:   3,452,723 ns/iter (+/- 108,821)
test es2017_cjs  ... bench:   3,545,935 ns/iter (+/- 143,864)
test es2018      ... bench:   3,501,772 ns/iter (+/- 117,469)
test es2018_cjs  ... bench:   3,614,023 ns/iter (+/- 299,849)
test es2019      ... bench:   3,394,230 ns/iter (+/- 79,776)
test es2019_cjs  ... bench:   3,752,338 ns/iter (+/- 102,233)
test es2020      ... bench:   3,062,410 ns/iter (+/- 98,368)
test es2020_cjs  ... bench:   3,182,953 ns/iter (+/- 417,189)
test es3         ... bench:   5,722,968 ns/iter (+/- 767,479)
test es3_cjs     ... bench:   7,048,859 ns/iter (+/- 2,742,203)
test es5         ... bench:   5,786,076 ns/iter (+/- 842,570)
test es5_cjs     ... bench:   5,966,745 ns/iter (+/- 816,473)

```

### Added more detailed benchmarks

```
test base_parser                  ... bench:     874,004 ns/iter (+/- 68,323)
test base_tr_clone                ... bench:      79,803 ns/iter (+/- 4,855)
test base_tr_es2016               ... bench:      72,714 ns/iter (+/- 2,526)
test base_tr_es2017               ... bench:      73,638 ns/iter (+/- 3,353)
test base_tr_es2018               ... bench:     156,399 ns/iter (+/- 11,838)
test base_tr_es2020               ... bench:     430,490 ns/iter (+/- 31,165)
test base_tr_fixer                ... bench:     167,136 ns/iter (+/- 10,122)
test base_tr_resolver_and_hygiene ... bench:     476,340 ns/iter (+/- 25,743)
test base_tr_typescript_strip     ... bench:     475,823 ns/iter (+/- 33,000)
test codegen_es2015               ... bench:     135,309 ns/iter (+/- 7,256)
test codegen_es2016               ... bench:     134,924 ns/iter (+/- 3,969)
test codegen_es2017               ... bench:     135,018 ns/iter (+/- 8,869)
test codegen_es2018               ... bench:     135,750 ns/iter (+/- 3,029)
test codegen_es2019               ... bench:     136,545 ns/iter (+/- 6,919)
test codegen_es2020               ... bench:     135,898 ns/iter (+/- 7,693)
test codegen_es3                  ... bench:     136,498 ns/iter (+/- 25,146)
test codegen_es5                  ... bench:     135,039 ns/iter (+/- 3,743)
test config_for_file              ... bench:      19,409 ns/iter (+/- 424)
test config_transforms            ... bench:   2,522,729 ns/iter (+/- 208,957)
test full_es2015                  ... bench:   5,283,123 ns/iter (+/- 193,428)
test full_es2016                  ... bench:   3,459,236 ns/iter (+/- 81,541)
test full_es2017                  ... bench:   3,461,171 ns/iter (+/- 86,303)
test full_es2018                  ... bench:   3,793,133 ns/iter (+/- 250,615)
test full_es2019                  ... bench:   3,606,445 ns/iter (+/- 473,261)
test full_es2020                  ... bench:   3,930,393 ns/iter (+/- 494,299)
test full_es3                     ... bench:   5,880,923 ns/iter (+/- 1,135,610)
test full_es5                     ... bench:   6,258,212 ns/iter (+/- 714,165)
```



### Moved benchmarks into the transforms

```
test base                               ... bench:     133,380 ns/iter (+/- 4,998) = 81 MB/s
test es2015                             ... FAILED
test es2015_arrow                       ... bench:     117,107 ns/iter (+/- 68,309) = 93 MB/s
test es2015_block_scoped_fn             ... bench:     216,823 ns/iter (+/- 156,713) = 50 MB/s
test es2015_block_scoping               ... bench:     292,041 ns/iter (+/- 130,519) = 37 MB/s
test es2015_classes                     ... FAILED
test es2015_computed_props              ... bench:     192,459 ns/iter (+/- 133,861) = 56 MB/s
test es2015_destructuring               ... bench:     440,398 ns/iter (+/- 261,789) = 24 MB/s
test es2015_duplicate_keys              ... bench:     193,026 ns/iter (+/- 63,876) = 56 MB/s
test es2015_fn_name                     ... bench:     188,594 ns/iter (+/- 29,205) = 57 MB/s
test es2015_for_of                      ... bench:      88,993 ns/iter (+/- 4,940) = 122 MB/s
test es2015_instanceof                  ... bench:     102,998 ns/iter (+/- 20,044) = 105 MB/s
test es2015_parameters                  ... bench:     174,503 ns/iter (+/- 887) = 62 MB/s
test es2015_shorthand_property          ... bench:     138,171 ns/iter (+/- 50,876) = 78 MB/s
test es2015_spread                      ... bench:     155,841 ns/iter (+/- 68,626) = 69 MB/s
test es2015_sticky_regex                ... bench:     173,401 ns/iter (+/- 113,243) = 62 MB/s
test es2015_typeof_symbol               ... bench:     118,730 ns/iter (+/- 52,377) = 91 MB/s
test es2016                             ... bench:      94,602 ns/iter (+/- 11,932) = 115 MB/s
test es2016_exponentation               ... bench:      91,149 ns/iter (+/- 16,240) = 119 MB/s
test es2017                             ... bench:      92,507 ns/iter (+/- 10,789) = 117 MB/s
test es2017_async_to_generator          ... bench:      93,964 ns/iter (+/- 46,257) = 115 MB/s
test es2018                             ... bench:     192,782 ns/iter (+/- 4,532) = 56 MB/s
test es2018_object_rest_spread          ... bench:     127,814 ns/iter (+/- 8,123) = 85 MB/s
test es2020                             ... bench:     562,368 ns/iter (+/- 332,788) = 19 MB/s
test es2020_class_properties            ... bench:     368,667 ns/iter (+/- 179,805) = 29 MB/s
test es2020_nullish_coalescing          ... bench:     218,601 ns/iter (+/- 60,245) = 49 MB/s
test es2020_optional_chaining           ... bench:     247,838 ns/iter (+/- 175,762) = 43 MB/s
test es2020_typescript_class_properties ... bench:     249,875 ns/iter (+/- 214,276) = 43 MB/s
test es3                                ... bench:     317,378 ns/iter (+/- 60,561) = 34 MB/s
```

### After adding fast paths to bullish coalescing and optional chaining

```
test base                               ... bench:     151,551 ns/iter (+/- 3,379) = 71 MB/s
test es2015                             ... FAILED
test es2015_arrow                       ... bench:     130,936 ns/iter (+/- 27,337) = 83 MB/s
test es2015_block_scoped_fn             ... bench:     134,872 ns/iter (+/- 2,936) = 80 MB/s
test es2015_block_scoping               ... bench:     135,010 ns/iter (+/- 9,195) = 80 MB/s
test es2015_classes                     ... FAILED
test es2015_computed_props              ... bench:      73,758 ns/iter (+/- 1,012) = 147 MB/s
test es2015_destructuring               ... bench:     280,959 ns/iter (+/- 19,996) = 38 MB/s
test es2015_duplicate_keys              ... bench:     145,697 ns/iter (+/- 5,232) = 74 MB/s
test es2015_fn_name                     ... bench:     135,559 ns/iter (+/- 4,034) = 80 MB/s
test es2015_for_of                      ... bench:      76,054 ns/iter (+/- 8,517) = 143 MB/s
test es2015_instanceof                  ... bench:      95,023 ns/iter (+/- 1,653) = 114 MB/s
test es2015_parameters                  ... bench:     143,098 ns/iter (+/- 8,742) = 76 MB/s
test es2015_shorthand_property          ... bench:     128,260 ns/iter (+/- 2,484) = 84 MB/s
test es2015_spread                      ... bench:     143,273 ns/iter (+/- 6,612) = 76 MB/s
test es2015_sticky_regex                ... bench:     143,377 ns/iter (+/- 1,170) = 75 MB/s
test es2015_typeof_symbol               ... bench:      97,128 ns/iter (+/- 1,423) = 112 MB/s
test es2016                             ... bench:      73,527 ns/iter (+/- 3,698) = 148 MB/s
test es2016_exponentation               ... bench:      72,688 ns/iter (+/- 2,912) = 149 MB/s
test es2017                             ... bench:      75,584 ns/iter (+/- 612) = 144 MB/s
test es2017_async_to_generator          ... bench:      75,952 ns/iter (+/- 576) = 143 MB/s
test es2018                             ... bench:     158,010 ns/iter (+/- 5,109) = 68 MB/s
test es2018_object_rest_spread          ... bench:     103,734 ns/iter (+/- 3,434) = 104 MB/s
test es2020                             ... bench:     298,838 ns/iter (+/- 2,588) = 36 MB/s
test es2020_class_properties            ... bench:     223,421 ns/iter (+/- 2,494) = 48 MB/s
test es2020_nullish_coalescing          ... bench:      73,575 ns/iter (+/- 1,581) = 148 MB/s
test es2020_optional_chaining           ... bench:     122,034 ns/iter (+/- 3,676) = 89 MB/s
test es2020_typescript_class_properties ... bench:     206,590 ns/iter (+/- 3,381) = 52 MB/s
test es3                                ... bench:     282,750 ns/iter (+/- 1,815) = 38 MB/s
```