@kdy1 2020-08-26
`#[fast_path]`

### For arrow

```
test base                               ... bench:     136,777 ns/iter (+/- 11,066) = 79 MB/s
test es2015                             ... FAILED
test es2015_arrow                       ... bench:     108,107 ns/iter (+/- 2,733) = 100 MB/s
test es2015_block_scoped_fn             ... bench:     134,961 ns/iter (+/- 8,728) = 80 MB/s
test es2015_block_scoping               ... bench:     144,212 ns/iter (+/- 599) = 75 MB/s
test es2015_classes                     ... FAILED
test es2015_computed_props              ... bench:      73,056 ns/iter (+/- 6,029) = 149 MB/s
test es2015_destructuring               ... bench:     306,134 ns/iter (+/- 7,927) = 35 MB/s
test es2015_duplicate_keys              ... bench:     156,429 ns/iter (+/- 613) = 69 MB/s
test es2015_fn_name                     ... bench:     135,204 ns/iter (+/- 3,364) = 80 MB/s
test es2015_for_of                      ... bench:      80,775 ns/iter (+/- 6,479) = 134 MB/s
test es2015_instanceof                  ... bench:     102,260 ns/iter (+/- 621) = 106 MB/s
test es2015_parameters                  ... bench:     153,001 ns/iter (+/- 804) = 71 MB/s
test es2015_shorthand_property          ... bench:     134,859 ns/iter (+/- 36,520) = 80 MB/s
test es2015_spread                      ... bench:     153,495 ns/iter (+/- 1,280) = 70 MB/s
test es2015_sticky_regex                ... bench:     152,677 ns/iter (+/- 818) = 71 MB/s
test es2015_typeof_symbol               ... bench:     102,544 ns/iter (+/- 9,114) = 106 MB/s
test es2016                             ... bench:      78,506 ns/iter (+/- 1,267) = 138 MB/s
test es2016_exponentation               ... bench:      78,716 ns/iter (+/- 1,649) = 138 MB/s
test es2017                             ... bench:      73,651 ns/iter (+/- 1,211) = 147 MB/s
test es2017_async_to_generator          ... bench:      80,133 ns/iter (+/- 345) = 135 MB/s
test es2018                             ... bench:     169,567 ns/iter (+/- 990) = 64 MB/s
test es2018_object_rest_spread          ... bench:     112,552 ns/iter (+/- 660) = 96 MB/s
test es2020                             ... bench:     297,361 ns/iter (+/- 11,942) = 36 MB/s
test es2020_class_properties            ... bench:     239,785 ns/iter (+/- 19,879) = 45 MB/s
test es2020_nullish_coalescing          ... bench:      79,287 ns/iter (+/- 338) = 137 MB/s
test es2020_optional_chaining           ... bench:     132,121 ns/iter (+/- 4,737) = 82 MB/s
test es2020_typescript_class_properties ... bench:     221,500 ns/iter (+/- 822) = 49 MB/s
test es3                                ... bench:     302,563 ns/iter (+/- 1,833) = 35 MB/s
```