@kdy1 2020-08-26
# Current status

```
test base                               ... bench:      66,029 ns/iter (+/- 1,051) = 164 MB/s
test common_reserved_word               ... bench:     125,766 ns/iter (+/- 1,934) = 86 MB/s
test es2015                             ... FAILED
test es2015_arrow                       ... bench:     107,558 ns/iter (+/- 790) = 101 MB/s
test es2015_block_scoped_fn             ... bench:     135,541 ns/iter (+/- 3,505) = 80 MB/s
test es2015_block_scoping               ... bench:     133,153 ns/iter (+/- 46,289) = 81 MB/s
test es2015_classes                     ... FAILED
test es2015_computed_props              ... bench:      72,620 ns/iter (+/- 3,328) = 149 MB/s
test es2015_destructuring               ... bench:     285,527 ns/iter (+/- 3,165) = 38 MB/s
test es2015_duplicate_keys              ... bench:     145,093 ns/iter (+/- 8,114) = 75 MB/s
test es2015_fn_name                     ... bench:     134,761 ns/iter (+/- 3,348) = 80 MB/s
test es2015_for_of                      ... bench:      76,804 ns/iter (+/- 1,770) = 141 MB/s
test es2015_instanceof                  ... bench:      94,329 ns/iter (+/- 2,103) = 115 MB/s
test es2015_parameters                  ... bench:     144,146 ns/iter (+/- 4,581) = 75 MB/s
test es2015_shorthand_property          ... bench:     129,914 ns/iter (+/- 2,794) = 83 MB/s
test es2015_spread                      ... bench:     148,342 ns/iter (+/- 4,144) = 73 MB/s
test es2015_sticky_regex                ... bench:     148,244 ns/iter (+/- 3,523) = 73 MB/s
test es2015_typeof_symbol               ... bench:      99,212 ns/iter (+/- 3,759) = 109 MB/s
test es2016                             ... bench:      78,380 ns/iter (+/- 3,649) = 138 MB/s
test es2016_exponentation               ... bench:      81,985 ns/iter (+/- 475) = 132 MB/s
test es2017                             ... bench:     101,850 ns/iter (+/- 23,006) = 106 MB/s
test es2017_async_to_generator          ... bench:      90,761 ns/iter (+/- 485) = 119 MB/s
test es2018                             ... bench:     119,925 ns/iter (+/- 575) = 90 MB/s
test es2018_object_rest_spread          ... bench:     126,204 ns/iter (+/- 15,086) = 86 MB/s
test es2018_optional_catch_binding      ... bench:      92,593 ns/iter (+/- 444) = 117 MB/s
test es2020                             ... bench:     332,526 ns/iter (+/- 2,800) = 32 MB/s
test es2020_class_properties            ... bench:     245,324 ns/iter (+/- 1,737) = 44 MB/s
test es2020_nullish_coalescing          ... bench:      93,031 ns/iter (+/- 583) = 117 MB/s
test es2020_optional_chaining           ... bench:     155,871 ns/iter (+/- 2,740) = 69 MB/s
test es2020_typescript_class_properties ... bench:     254,524 ns/iter (+/- 2,021) = 42 MB/s
test es3                                ... bench:     348,161 ns/iter (+/- 2,200) = 31 MB/s

```