@Brooooooklyn 2020-08-26
Latest benchmark from https://github.com/kdy1/swc `perf` branch

## Sync

```bash
@swc-node/core x 236 ops/sec ±2.16% (85 runs sampled)
esbuild x 49.42 ops/sec ±1.77% (63 runs sampled)
typescript x 27.75 ops/sec ±10.04% (52 runs sampled)
babel x 21.78 ops/sec ±11.33% (45 runs sampled)
Transform rxjs/AjaxObservable.ts benchmark bench suite: Fastest is @swc-node/core
```
## Single thread async

```bash
@swc-node/core x 182 ops/sec ±3.04% (77 runs sampled)
esbuild x 464 ops/sec ±2.02% (76 runs sampled)
Transform rxjs/AjaxObservable.ts async benchmark bench suite: Fastest is esbuild
```

## Parallel async

```bash
@swc-node/core x 597 ops/sec ±1.15% (76 runs sampled)
esbuild x 1,012 ops/sec ±2.01% (67 runs sampled)
Transform rxjs/AjaxObservable.ts parallel benchmark bench suite: Fastest is esbuild
```