@kdy1 2020-08-26
# Fix for typescript::strip
There was a design flaw in typescript stripper, and I fixed it.

```
test base                               ... bench:      69,063 ns/iter (+/- 435) = 157 MB/s
test common_typescript                  ... bench:     150,591 ns/iter (+/- 2,181) = 72 MB/s
```

base: time for `Module.clone()`